+++ code/test/sixsq/nuvla/server/resources/session_password_lifecycle_test.clj
@@ -2,7 +2,7 @@
   (:require
     [clojure.data.json :as json]
     [clojure.string :as str]
-    [clojure.test :refer [deftest is use-fixtures]]
+    [clojure.test :refer [deftest is use-fixtures testing]]
     [peridot.core :refer [content-type header request session]]
     [postal.core :as postal]
     [sixsq.nuvla.auth.utils :as auth]
@@ -279,7 +279,7 @@
 
     ))
 
-(deftest claim-account-test
+(deftest switch-group-lifecycle-test
 
   (let [app                (ltu/ring-app)
         session-json       (content-type (session app) "application/json")
@@ -470,7 +470,7 @@
 
         ))))
 
-(deftest get-peers-test
+(deftest get-peers-lifecycle-test
   (let [app                (ltu/ring-app)
         session-json       (content-type (session app) "application/json")
         session-user       (header session-json authn-info-header "user group/nuvla-user")
@@ -528,3 +528,157 @@
             (ltu/body)
             (= {})
             (is "Get peers body should be empty"))))))
+
+
+(deftest get-groups-lifecycle-test
+  (let [app              (ltu/ring-app)
+        session-json     (content-type (session app) "application/json")
+        session-anon     (header session-json authn-info-header "user/unknown user/unknown group/nuvla-anon")
+        session-admin    (header session-json authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
+        user-id          (create-user session-admin
+                                      :username "tarzan"
+                                      :password "TarzanTarzan-0"
+                                      :activated? true
+                                      :email "tarzan@example.org")
+        session-user     (header session-json authn-info-header (str user-id user-id " group/nuvla-user group/nuvla-anon"))
+        session-group-a  (header session-json authn-info-header "user/x group/a user/x group/nuvla-user group/nuvla-anon group/a")
+        session-group-b  (header session-json authn-info-header "user/x group/b user/x group/nuvla-user group/nuvla-anon group/b")
+        href             (str st/resource-type "/password")
+
+        grp-base-uri     (str p/service-context "group")
+        valid-create-grp (fn [group-id] {:template {:href             "group-template/generic"
+                                                    :group-identifier group-id
+                                                    :name             (str "Group " group-id)
+                                                    :description (str "Group " group-id " description") }})]
+
+    (-> session-admin
+        (request grp-base-uri
+                 :request-method :post
+                 :body (json/write-str (valid-create-grp "a")))
+        (ltu/body->edn)
+        (ltu/is-status 201))
+    (-> session-group-a
+        (request grp-base-uri
+                 :request-method :post
+                 :body (json/write-str (valid-create-grp "b")))
+        (ltu/body->edn)
+        (ltu/is-status 201))
+    (-> session-group-a
+        (request grp-base-uri
+                 :request-method :post
+                 :body (json/write-str (valid-create-grp "b1")))
+        (ltu/body->edn)
+        (ltu/is-status 201))
+    (-> session-group-b
+        (request grp-base-uri
+                 :request-method :post
+                 :body (json/write-str (valid-create-grp "c")))
+        (ltu/body->edn)
+        (ltu/is-status 201))
+
+    ; anonymous create must succeed
+    (let [resp            (-> session-anon
+                              (request base-uri
+                                       :request-method :post
+                                       :body (json/write-str {:template {:href     href
+                                                                         :username "tarzan"
+                                                                         :password "TarzanTarzan-0"}}))
+                              (ltu/body->edn)
+                              (ltu/is-set-cookie)
+                              (ltu/is-status 201))
+          id              (ltu/body-resource-id resp)
+          abs-uri         (ltu/location-url resp)
+          session-with-id (header session-json authn-info-header (str user-id user-id " group/nuvla-user group/nuvla-anon " id))]
+      ; user should be able to see session with session role
+      (-> session-with-id
+          (request abs-uri)
+          (ltu/body->edn)
+          (ltu/is-status 200)
+          (ltu/is-id id)
+          (ltu/is-operation-present :delete)
+          (ltu/is-operation-absent :edit)
+          (ltu/is-operation-absent :switch-group)
+          (ltu/is-operation-present :get-peers)
+          (ltu/is-operation-present :get-groups))
+
+      ; check contents of session
+      (let [get-groups-url (-> session-user
+                               (header authn-info-header (str user-id " " user-id " group/nuvla-user group/nuvla-anon " id))
+                               (request abs-uri)
+                               (ltu/body->edn)
+                               (ltu/get-op-url :get-groups))]
+
+        (testing "user who is not in any group should get empty list of groups"
+          (-> session-with-id
+              (request get-groups-url)
+              (ltu/body->edn)
+              (ltu/is-status 200)
+              (ltu/body)
+              (= [])
+              (is "Get groups body should have no childs")))
+
+
+        (testing
+          "when user is part of a group he should get
+           the subgroups"
+          (-> session-admin
+              (request (str p/service-context "group/b")
+                       :request-method :put
+                       :body (json/write-str {:users [user-id]}))
+              (ltu/body->edn)
+              (ltu/is-status 200))
+          (-> session-with-id
+              (request get-groups-url)
+              (ltu/body->edn)
+              (ltu/is-status 200)
+              (ltu/body)
+              (= [{:children    [{:description "Group c description"
+                                  :id          "group/c"
+                                  :name        "Group c"}]
+                   :description "Group b description"
+                   :id          "group/b"
+                   :name        "Group b"}])
+              (is "User get group/b and subgroup group/c")))
+
+        (testing
+          "when user is part of a root group he should get
+           the full group hierarchy and group/b is not duplicated"
+          (-> session-admin
+              (request (str p/service-context "group/a")
+                       :request-method :put
+                       :body (json/write-str {:users [user-id]}))
+              (ltu/body->edn)
+              (ltu/is-status 200))
+          (-> session-admin
+              (request grp-base-uri
+                       :request-method :post
+                       :body (json/write-str (valid-create-grp "z")))
+              (ltu/body->edn)
+              (ltu/is-status 201))
+          (-> session-admin
+              (request (str p/service-context "group/z")
+                       :request-method :put
+                       :body (json/write-str {:users [user-id]}))
+              (ltu/body->edn)
+              (ltu/is-status 200))
+          (-> session-with-id
+              (request get-groups-url)
+              (ltu/body->edn)
+              (ltu/is-status 200)
+              (ltu/body)
+              (= [{:children    [{:children    [{:description "Group c description"
+                                                 :id          "group/c"
+                                                 :name        "Group c"}]
+                                  :description "Group b description"
+                                  :id          "group/b"
+                                  :name        "Group b"}
+                                 {:description "Group b1 description"
+                                  :id          "group/b1"
+                                  :name        "Group b1"}]
+                   :description "Group a description"
+                   :id          "group/a"
+                   :name        "Group a"}
+                  {:description "Group z description"
+                   :id          "group/z"
+                   :name        "Group z"}])
+              (is "Get groups body should contain tree of groups")))))))